# api/openapi.yaml
openapi: 3.0.3
info:
  title: recommend4me
  version: 0.1.0
  description: |
    A local recommender of works — games, books — collected from sites by plugins. The user grades
    some of them; a model learns the order of the grades and puts every other work on the user's
    own 0..10 scale.

    An error is answered as application/problem+json (the Problem schema): its "detail" says, in
    the user's language, what went wrong.
...
paths:
...
  /api/items/{source}/{item}/corrections/facets:
    put:
      tags: [corrections]
      operationId: correctFacet
      summary: Adds a value of a facet to the item, or takes one the site gives away
      parameters:
        - $ref: '#/components/parameters/Source'
        - $ref: '#/components/parameters/Item'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FacetCorrection'
      responses:
        '200':
          description: The item after the correction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemDetails'
    delete:
      tags: [corrections]
      operationId: uncorrectFacet
      summary: Takes back a correction of a facet value
      parameters:
        - $ref: '#/components/parameters/Source'
        - $ref: '#/components/parameters/Item'
        - name: facet
          in: query
          required: true
          schema:
            type: string
        - name: key
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The item after the correction is taken back
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemDetails'
  /api/items/{source}/{item}/corrections/fields/{field}:
    put:
      tags: [corrections]
      operationId: correctField
      summary: Overrides a field of the item ("title", "text:<key>", "number:<key>")
      parameters:
        - $ref: '#/components/parameters/Source'
        - $ref: '#/components/parameters/Item'
        - $ref: '#/components/parameters/Field'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FieldCorrection'
      responses:
        '200':
          description: The item after the correction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemDetails'
    delete:
      tags: [corrections]
      operationId: uncorrectField
      parameters:
        - $ref: '#/components/parameters/Source'
        - $ref: '#/components/parameters/Item'
        - $ref: '#/components/parameters/Field'
      responses:
        '200':
          description: The item with the site's value of the field again
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemDetails'
  /api/items/{source}/{item}/links:
    post:
      tags: [corrections]
      operationId: linkItem
      summary: Says the item and another one (of any source of the type) are one work
      parameters:
        - $ref: '#/components/parameters/Source'
        - $ref: '#/components/parameters/Item'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemRef'
      responses:
        '200':
          description: The item after the link
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemDetails'
        '400':
          description: The other item is of another content type, or unknown
    delete:
      tags: [corrections]
      operationId: unlinkItem
      parameters:
        - $ref: '#/components/parameters/Source'
        - $ref: '#/components/parameters/Item'
        - name: otherSource
          in: query
          required: true
          schema:
            type: string
        - name: otherItem
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The item after the link is taken back
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemDetails'
...
components:
  parameters:
...
    Source:
      name: source
      in: path
      required: true
      schema:
        type: string
    Item:
      name: item
      in: path
      required: true
      schema:
        type: string
...
    Field:
      name: field
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    Grade:
      type: integer
      minimum: 1
      maximum: 5
      description: |
        The user's grade, an order and not a number: 1 - do not like it, 2 - can be played (read),
        3 - liked it on the whole, 4 - very good, 5 - give me more.
...
    FacetValueInfo:
      type: object
      required: [key, name]
      properties:
        key:
          type: string
        name:
          type: string
        corrected:
          type: string
          enum: [ADDED, REMOVED]
          description: The user's correction of the site's value; a REMOVED value is shown crossed out
    ItemFacet:
      type: object
      required: [facet, label, values]
      properties:
        facet:
          type: string
          description: The facet's key within its source
        label:
          type: string
        values:
          type: array
          items:
            $ref: '#/components/schemas/FacetValueInfo'
    Prediction:
      type: object
      required: [score]
      properties:
        score:
          type: number
          format: double
          description: |
            The work's place on the user's own scale: the centre of the lowest grade is 0, of the
            highest 10; works beyond the extreme grades go beyond 0..10
    ItemRef:
      type: object
      required: [source, item]
      properties:
        source:
          type: string
        item:
          type: string
    LinkedItem:
      type: object
      required: [source, item, title, url]
      properties:
        source:
          type: string
        item:
          type: string
        title:
          type: string
        url:
          type: string
    ItemSummary:
      type: object
      required: [source, item, title, version, url, updatedAt, facets, numbers, hasCover, pictureCount, linked, work]
      properties:
        work:
          type: string
          description: |
            The work the item is of, as "<source>/<item>" of the item that stands for it in the
            list; the item's own key when it is linked to no other
        source:
          type: string
        item:
          type: string
        title:
          type: string
        version:
          type: string
        url:
          type: string
        updatedAt:
          type: string
          format: date-time
        facets:
          type: array
          description: The facets shown on a card
          items:
            $ref: '#/components/schemas/ItemFacet'
        numbers:
          type: object
          description: By the number's key within the item's source (SourceInfo.numbers[].key), corrections applied
          additionalProperties:
            type: number
            format: double
        hasCover:
          type: boolean
        pictureCount:
          type: integer
        grade:
          description: The grade of the work's current version - given to this item or to another item of the same work
          allOf:
            - $ref: '#/components/schemas/Grade'
        previousGrade:
          $ref: '#/components/schemas/Grade'
        previousGradeVersion:
          type: string
        prediction:
          $ref: '#/components/schemas/Prediction'
        searchMatch:
          $ref: '#/components/schemas/SearchMatch'
        signals:
          type: object
          description: The user's own actions on the site (liked, on a shelf)
          additionalProperties:
            type: string
        linked:
          type: array
          description: The other items that are this same work
          items:
            $ref: '#/components/schemas/LinkedItem'
    SearchMatch:
      type: object
      required: [field, byMeaning, text, highlights]
      properties:
        field:
          type: string
          description: The name of the field it was found in, as the user is shown it
        byMeaning:
          type: boolean
          description: Found by meaning alone; the text is then the phrase of the field nearest by meaning
        text:
          type: string
        highlights:
          type: array
          items:
            $ref: '#/components/schemas/TextRange'
    TextRange:
      type: object
      required: [start, end]
      properties:
        start:
          type: integer
        end:
          type: integer
...
    TextValue:
      type: object
      required: [key, label, content, spoiler, corrected]
      properties:
        key:
          type: string
        label:
          type: string
        content:
          type: string
        spoiler:
          type: boolean
        corrected:
          type: boolean
    NumberValue:
      type: object
      required: [key, label, value, corrected]
      properties:
        key:
          type: string
        label:
          type: string
        value:
          type: number
          format: double
        corrected:
          type: boolean
    PartInfo:
      type: object
      required: [partId, position, hasText]
      properties:
        partId:
          type: string
        position:
          type: integer
        title:
          type: string
        hasText:
          type: boolean
        publishedAt:
          type: string
          format: date-time
...
    ItemDetails:
      type: object
      required: [summary, allFacets, texts, allNumbers, explanation, ratings, pictures, reviews, parts, pictureMatches, reviewMatches, titleCorrected, canRefresh]
      properties:
        summary:
          $ref: '#/components/schemas/ItemSummary'
        titleCorrected:
          type: boolean
        allFacets:
          type: array
          description: Every facet of the item, corrections included
          items:
            $ref: '#/components/schemas/ItemFacet'
        texts:
          type: array
          items:
            $ref: '#/components/schemas/TextValue'
        allNumbers:
          type: array
          items:
            $ref: '#/components/schemas/NumberValue'
        explanation:
          type: array
          description: Features that pushed the prediction most, strongest first
          items:
            $ref: '#/components/schemas/FeatureContribution'
        ratings:
          type: array
          description: The grades of every item of the work, newest first
          items:
            $ref: '#/components/schemas/RatingRecord'
        reviews:
          type: array
          description: The reviews that move the work most either way, strongest first; or the newest
          items:
            $ref: '#/components/schemas/ReviewInfo'
        reviewCount:
          type: integer
        pictures:
          type: array
          items:
            $ref: '#/components/schemas/PictureInfo'
        parts:
          type: array
          items:
            $ref: '#/components/schemas/PartInfo'
        pictureMatches:
          type: array
          items:
            $ref: '#/components/schemas/PictureMatch'
        reviewMatches:
          type: array
          items:
            $ref: '#/components/schemas/ReviewMatch'
        canRefresh:
          type: boolean
          description: The source can download the item's page again
...
    PictureMatch:
      type: object
      required: [position, markSource, markItem, markTitle, markPosition, mark, strength]
      properties:
        position:
          type: integer
        markSource:
          type: string
        markItem:
          type: string
        markTitle:
          type: string
        markPosition:
          type: integer
        verdict:
          type: integer
        mark:
          type: integer
        strength:
          type: number
          format: double
          description: |
            How surprising the likeness is: -ln of the chance that as many random pictures of the
            catalogue would come as close; chance gives about 1, 3 one work in twenty, 5 one in 150
    ReviewMatch:
      type: object
      required: [reviewId, content, markSource, markItem, markReviewId, markTitle, markContent, mark, strength]
      properties:
        reviewId:
          type: string
        content:
          type: string
        markSource:
          type: string
        markItem:
          type: string
        markReviewId:
          type: string
        markTitle:
          type: string
        markContent:
          type: string
        verdict:
          type: integer
        mark:
          type: integer
        strength:
          type: number
          format: double
    ReviewInfo:
      type: object
      required: [reviewId, content]
      properties:
        reviewId:
          type: string
        author:
          type: string
        stars:
          type: number
          format: double
        content:
          type: string
        postedAt:
          type: string
          format: date-time
        influence:
          type: number
          format: double
          description: Points of the 0..10 scale the review moves the work, left out of the reviews' set and of their likeness to the marked ones
        mark:
          type: integer
...
    PictureInfo:
      type: object
      required: [position, hasFull, analyzed]
      properties:
        position:
          type: integer
          description: 0 the cover, 1.. the screenshots
        hasFull:
          type: boolean
        analyzed:
          type: boolean
        mark:
          type: integer
        influence:
          type: number
          format: double
        error:
          type: string
    RatingRecord:
      type: object
      required: [source, item, version, grade, ratedAt, current]
      properties:
        source:
          type: string
        item:
          type: string
          description: The item of the work graded - this one or one linked to it
        current:
          type: boolean
          description: The grade is of the item's current version
        version:
          type: string
        grade:
          $ref: '#/components/schemas/Grade'
        ratedAt:
          type: string
          format: date-time
    FeatureContribution:
      type: object
      required: [feature, label, contribution, present]
      properties:
        feature:
          type: string
        label:
          type: string
        contribution:
          type: number
          format: double
          description: Points of the 0..10 scale the feature moves the work
        present:
          type: boolean
...
    FacetCorrection:
      type: object
      required: [facet, added]
      properties:
        facet:
          type: string
          description: The facet's key within the item's source
        key:
          type: string
          description: |
            The value's key; when none, the value is found by its name among the facet's values
            (any case), and a new one gets the name in lower case as its key
        name:
          type: string
          description: The name of a value the user makes up; required without a key
        added:
          type: boolean
          description: true - the value is added; false - the site's value is taken away
    FieldCorrection:
      type: object
      required: [content]
      properties:
        content:
          type: string
...